requirements.txt
openai
pinecone
cohere
tiktoken
streamlit
python-dotenv
